Add TaskGenerators and remove Masters
Extending on the theme of independent tasks, this commit introduces task
generators to replace the "meta" nodes.  Task generators are python
generator functions that yield lists of independent tasks are fed back
their results.  This design allows to run this generator on a central
machine and distribute the tasks to any number of configured remote
compute resources with very loose coupling.  It's not clear to me yet,
that python generator functions are sufficient beyond the PoC stage, but
it works for now.

As part of this development the previous executors are split into
execution contexts, concerned with just list of parallel tasks and
executors themselves which manage some compute resource (i.e. a process
pool) and let the execution contexts run within.

We've also lost the ability to run jobs and tasks completely in the
background, because I haven't figured it out completely yet.