TrainingContainer: Allow to customize HDF keys
Allows to change in the location from hdf file where training data is read from by TrainingContainer.include_job.

This also adds a Vasp specific work around to read the smeared energy from the file instead of the generic energy_pot (which is the sigma to zero extrapolated energy), which previously resulted in inconsistent energies and forces.